some changes 29.05.17

+++ src/main/java/com/springapp/market/controller/ProductController.java
@@ -61,13 +61,13 @@ public class ProductController {
     }
 
     @RequestMapping("/remove_product/{id}")
-    public String removeProduct(@PathVariable("id") int id) {
+    public String removeProduct(@PathVariable("id") long id) {
         this.productService.removeProduct(id);
         return "redirect:/products";
     }
 
     @RequestMapping("edit_product/{id}")
-    public String editProduct(@PathVariable("id") int id, Model model) {
+    public String editProduct(@PathVariable("id") long id, Model model) {
         model.addAttribute("product", this.productService.getProductById(id));
         model.addAttribute("listProducts", this.productService.listProducts());
         model.addAttribute("listCompanies", this.companyService.listCompanies());
@@ -75,7 +75,7 @@ public class ProductController {
     }
 
     @RequestMapping("product_data/{id}")
-    public String productData(@PathVariable("id") int id, Model model) {
+    public String productData(@PathVariable("id") long id, Model model) {
         model.addAttribute("product", this.productService.getProductById(id));
         return "product_data";
     }
